+++ admin/includes/comments_table.php
@@ -53,7 +53,7 @@
 					</td>
 					<?php changeStatus();?>
 					<td><a href='comments.php?source=edit_comment&id=<?php echo $row['comment_id'];?>'>Edit</a></td>
-					<td><a href='comments.php?delete_comment=<?php echo $row['comment_id'];?>'>Delete</a></td>
+					<td><a onclick="return confirm('Are you sure you want to delete?')" href='comments.php?delete_comment=<?php echo $row['comment_id'];?>'>Delete</a></td>
 					<tr>
 					<?php
 				}
