+++ admin/includes/posts_table.php
@@ -17,12 +17,8 @@
 			</tr>
 		</thead>
 		<tbody>
-			<?php 
-			if(isset($_GET['delete_post'])){
-				deleteRowFromDb("posts", "delete_post", "post");
-				deleteRowFromDb("comments", "delete_post", "comments_post");
-				header("Location: posts.php");
-			}
+			<?php
+			deletePost();
 			$current_user = $_SESSION['user_name'];
 			$result = readFromdb("posts", " WHERE post_author='{$current_user}'");
 			while( $row = mysqli_fetch_assoc($result)){
@@ -30,7 +26,19 @@
 				<tr>
 				<td><?php echo $row['post_id'];?></td>
 				<td><?php echo $row['post_author'];?></td>
-				<td><a href="../index.php?source=show_post&id=<?php echo $row['post_id'];?>"><?php echo $row['post_title'];?></a></td>
+				<td>
+				<?php
+				if($row['post_status'] === 'published'){
+					?>
+					<a href="../index.php?source=show_post&id=<?php echo $row['post_id'];?>"><?php echo $row['post_title'];?></a>
+					<?php
+				} else {
+					?>
+					<?php echo $row['post_title'];?>
+					<?php
+				}
+				?>
+				</td>
 				<td><?php
 					$category_row = getCategoryTitleUsingCatId($row);
 					echo $category_row['cat_title'];
@@ -41,7 +49,7 @@
 				<td><?php echo countCommentsUsingId($row);?></td>
 				<td><?php echo $row['post_date'];?></td>
 				<td><a href='posts.php?source=edit_post&id=<?php echo $row['post_id'];?>'>Edit</a></td>
-				<td><a href='posts.php?delete_post=<?php echo $row['post_id'];?>'>Delete</a></td>
+				<td><a onclick="return confirm('Are you sure you want to delete?')" href='posts.php?delete_post=<?php echo $row['post_id'];?>'>Delete</a></td>
 				<tr>
 				<?php
 			}
